app/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
starlette==0.50.0
SQLAlchemy==2.0.46
pyodbc==5.1.0
python-dotenv==1.2.1
passlib==1.7.4
bcrypt==5.0.0
python-jose==3.5.0
cryptography==46.0.3
websockets==16.0
azure-cosmos==4.7.0
watchfiles==1.1.1
httptools==0.7.1
PyYAML==6.0.3
